+++ src/components/App/AppPages.js
@@ -1,4 +1,4 @@
-import React, {useContext} from 'react';
+import React, {useContext, useEffect, useState} from 'react';
 import Button from '@material-ui/core/Button';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
@@ -7,9 +7,8 @@ import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
 import Link from '@material-ui/core/Link';
-import {P_MENUS} from '../../utils/constants';
+import {P_MENUS, PRISMIC_SITEMEDIAS_DOC} from '../../utils/constants';
 import {UserContext} from '../../store/context/userContext';
-import bg from '../../assets/images/landingBg.png';
 import {useRenderCounts} from  '../../hooks/api-hooks';
 
 // import GridListTile from '@material-ui/core/GridListTile';
@@ -89,8 +88,15 @@ const cards = P_MENUS;
 
 export default function Album() {
   useRenderCounts('AppPages.js');  
+  const [bgimg, setBg] = useState('')
+  
   const classes = useStyles();
-  const {handleNav} = useContext(UserContext);
+  const {handleNav, cmsContents} = useContext(UserContext);
+  useEffect(() => {
+    if (cmsContents.data && cmsContents.data.hasOwnProperty(PRISMIC_SITEMEDIAS_DOC)) {
+      setBg(cmsContents.data[PRISMIC_SITEMEDIAS_DOC].edges[0].node.assetImage.url)
+    }
+  }, [cmsContents])
   
   return (
     <React.Fragment>      
@@ -98,7 +104,7 @@ export default function Album() {
         {/* Hero unit */}        
         <Container className={classes.cardGrid} maxWidth="md">
           {/* End hero unit */}
-          <Grid container spacing={1} style={{backgroundImage: `url(${bg})`, backgroundPosition: 'center',backgroundRepeat:'no-repeat', backgroundColor: '#fff'}}>
+          <Grid container spacing={1} style={{backgroundImage: `url(${bgimg})`, backgroundPosition: 'center',backgroundRepeat:'no-repeat', backgroundColor: '#fff'}}>
             {cards.map((card, indx) => (
               <Grid item key={`${indx}-${card.page}`} xs={6} sm={6} md={4} style={{padding:'0px', background:'transparent'}}>
                 <Card className={classes.card}>